+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2024-11-02
+
+### Added
+
+- `nil.to_list()` method
+
+### Changed
+
+- Dropped support for Python 3.11
+
 ## [0.4.1] - 2023-10-05
 
 ### Added